requirements.txt
feedparser==6.0.10
newspaper3k==0.2.8
EbookLib==0.18
PyGithub==2.1.1
google-generativeai>=0.8.0
PyYAML==6.0.1
beautifulsoup4==4.12.3
